+++ tailwind.config.ts
@@ -1,5 +1,6 @@
 import type { Config } from "tailwindcss";
 import animate from "tailwindcss-animate";
+import typography from "@tailwindcss/typography";
 
 export default {
 	darkMode: ["class"],
@@ -107,5 +108,5 @@ export default {
 				}
 			}
 	},
-	plugins: [animate],
+	plugins: [animate, typography],
 } satisfies Config;